Initial implementation of {@inject} for Soy templates. This syntax allows you to declare your injected parameters just like regular {@param} parameters, with strong types, and use them like regular variables ($ij no longer needed). This also means that you can inject proto types, since the type information needed to generate the getter / setter calls is available.
Internally the TemplateNode now maintains two lists of parameters, regular parameters and injected parameters. Injected parameters are not considered part of the template's calling signature, however they are required to conform to the usual rules of parameters (i.e. declaring a parameter and not using it is an error). You should only declare the injected parameters that you actually use.

For rendering in both Java and JS, the logic is pretty simple: If it's an injected parameter, get the value from 'ijData', otherwise get it from 'data'. Injected parameters will be runtime type checked just like regular strongly-typed parameters.

One open issue is that the template author cannot choose the name of an injected parameter, they can only use the name that is provided in the injected data. This means that if there is a naming conflict between a regular parameter and an injected parameter, the regular parameter must be renamed. At some future point we may look into adding parameter name aliases to address this issue.

This CL also simplifies the template parser grammar by refactoring it to not require as many lookaheads.
Also, the template parameter list is no longer nullable - it returns an empty list if there are no parameters.

Note: There are currently no integration tests in this CL yet.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=65454818