Soy templates: accept visibility="private" as a synonym for private="true".
Currently, Soy templates can be public or private; this is controlled by a "private" boolean-valued attribute on the template.

We would like to make more visibility options available to template authors, in particular package private. (See discussion in xrefed bug).

This CL makes it possible to support more visibility options in the future. In particular, internal Soy data structures currently use a boolean "isPrivate" flag to keep track of a template's visibility. This is inadequate when there are more than two visibility levels, so this CL changes the boolean flag to an enumeration. The enumeration has two values, PUBLIC and PRIVATE, with a third, PACKAGE, commented out so it is easy to add in the future.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=69180818