Add explicit autoescape attributes to Soy templates, and rename non-strict autoescape modes as deprecated.
The purpose of this change is to prepare Soy templates in [] for a
change in Soy's default autoescape mode to autoescape="strict", and for
deprecation of non-strict autescape modes.

Each Soy template is refactored as follows:
* Replace uses of deprecated autoescape modes ("false", "true",
"contextual") with their new names ("deprecated-noautoescape",
"deprecated-noncontextual", "deprecated-contextual", respectively).
Note that prior to this change, Soy has been updated to recognize both
old and new names as equivalent.
* Ensure that all templates specify an explicit autoescape mode, i.e.
don't rely on the implicit default of autoescape="true". This will
ensure that later changes in Soy to change default behavior will not
break existing templates.

This change is part of a refactoring to change Soy's default
autoescaping to "strict".  Strict autoescaping provides strong security
guarantees, and is the only Soy autoescape mode recommended by security
team (see []It is at this time widely used in
[] applications including G+ and Gmail. In addition, Soy's three
older autoescape modes, ranging from less secure to very insecure, are
being deprecated and correspondingly renamed.

The overall refactoring proceeds in a sequence of steps, each of which
is designed to introduce no modification in behavior in applications
using Soy. For details, see [] This change
is step (B) of that sequence.

More information: []announcement of this
refactoring at []
Tested:
[] --sample for global presubmit queue
[]All failing tests were failing before this change
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=70004395