Soy: wire CheckTemplateVisibility pass into SoyFileSetParser checking passes.
In [] I created the CheckTemplateVisibility pass to traverse the Soy AST and throw exceptions when a visibility="private" template was called from a call site where it was not visible. But that CL did not actually wire the pass in.

This CL wires the pass into the SoyFileSetParser's list of "checking passes".

Once this CL is submitted, the CheckTemplateVisibility pass will run whenever SoyFileSetParser#parse is executed and the parser is configured to run the checking passes. This will increase CPU utilization, even for file sets that have no visibility="private" templates, as it is another sequentially-run pass over the tree.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=70669905