Soy conformance: simplify proto definition: use oneof instead of enum.
The Soy conformance framework supports several types of conformance checks. Currently, this is encoded at the proto level using a single Requirement message with a type field.

This design forces the Requirement message to contain the union of all fields needed by all conformance checks. For example, the BannedCommandText check needs a "context" field to know what commands to look inside, but the other checks do not need this field.

This CL changes the Requirement to use the "oneof" feature. This allows each requirement type to contain only the fields relevant to it, and provides better enforcement that the relevant fields are actually set.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=72911850