Soy to Python CL #9. Design Doc: []
Support for ForeachNodes and local variable tracking.

The local variable tracking is something that's required by a few features, with foreach just being the first to implement it. Essentially we just keep a stack of the currently available local variables so that the expression translator will be able to reference them.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86990852