'Let' Nodes. Soy to Python Design Doc: []
Add support for 'Let' nodes. We made a decision fairly on to skip any support for non-strict templates, so let content nodes are one of the first places where that comes into effect. We mandate a contentKind on each let content node, and fail if it's not provided.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88549002