Introduce some core model classes for the bytecode compiler.
The key abstraction is an Expression, which has a type and can generate
bytecode to produce a value of that type on the runtime execution stack. This
cl introduces a number of low level ways for producing and manipulating
expressions.

* FieldRefs which can create expressions that read/load fields
* LocalVariable which can create expressions that read local variables
* MethodRefs which can create expressions that invoke methods
* SoyExpressions a subtype of expressions that represent soy runtime types,
  they support manipulations involving type coercions and boxing operations
* BytecodeUtils a set of utilities for transforming expressions.

In addition this introduces a few ancillary data structures
* MemoryClassLoader a classloader implementation that can load a class from a
  byte[]
* ClassData a simple tuple of a byte[] and some type information, this will be
  our compiler output format
* TypeInfo a wrapper around org.objectweb.asm.Type that is specialized for
  classes
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88754146