Soy JS backend - machinery for enumerating helper functions needed to do field access based on types.
https://docs.google.com[]document/d/1F9v14nI3E_BS9_h7uYC7RyqR3mvBF5l0PcaV7yAUA1U/edit?usp=sharing describes the approach.

This CL introduces two new classes but does not hook them up to anything.

HelperFunctions contains classes that allow the last part of the code-generator -- the one which writes output files -- to
(1) determine which goog.require and goog.provide declarations are needed in each generated source file
(2) how to generate helper functions with names that are meaningful in a debugger and with relevant comments.
(3) whether runtime type checks are needed inside a particular helper function

CollectTypeHelpersExprVisitor is a pass that should be run over all expressions in a SoyFileSet before TranslateToJsExprVisitor.  It collects information about the types of various expressions so that we can reuse helper functions as much as possible.
It pays particular attention to unions so that we can insert the minimal number* of runtime-type checks so as to allow JSCompiler to eliminate constructors as much as possible.

This will be followed by 3 more CLs.
(1) Which changes this to recognize that protobuf classes need field accessors,
    and that Safe HTML protos need value conversion.
(2) Which modified SoyToJS to create a helper function collector (based on a flag initially) and to
    create an output file with the helper functions.
(2) Which modifies TranslateToJsExprVisitor to run the collector and use its FieldAccessStrategies
    to generate code for field accesses and variable accesses.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88854556