Extract some functionality around basic operators from the EvalVisitor to a location where the jbcsrc backend can use it.
The jbcsrc backend will try to (as much as possible) directly implement these operators in generated code.  But for that to work we will need known static types, this library is intended to be a fallback for when we need runtime information to implement an operation.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88941850