Make @CompileTimeConstant parameters final
If a method's formal parameter is annotated with @CompileTimeConstant, the
method will always be invoked with an argument that is a static constant. If
the parameter itself is non-final, then it is a mutable reference to immutable
data. This is rarely useful, and can be confusing when trying to use the
parameter in a context that requires a compile-time constant.

Tested:
    [] --sample for global presubmit queue
    []    All failing tests were failing before this change
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=89069661