Support for Call nodes. Soy to Python Design Doc: []
There are a few things of note for the Call implementation. First of all we check content params for anything which isn't computable as a Py expression. Those are the only parts which can't be inlined, so we store those as params before the call.

Second, since params can be specified in a few different formats, the data provided to the callee is a combination of the data object provided, and any specified param nodes. These are merged into a new dictionary if both are present.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=89315627