Soy tests: replace nine-headed SharedTestUtils#parseSoy(Code|Files) with builder.
Currently, SharedTestUtils has 4 overloads of parseSoyCode and 6 overloads of parseSoyFiles. The overloads cover options like the syntax version to use and the compiler passes to run.

These options are reasonable, but they exist in a complex call hierarchy, with overloads delegating through multiple levels of overloads. This makes it hard to add new options and also for test authors to decide which is the right overload for them to use.

This CL replaces the overload hierarchy with SoyFileSetParserBuilder. Test authors create a builder, set any non-standard options, and then call parse().

The CL is unfortunately a lot of churn, but it's past time to pay the price to get a more scalable test setup.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=89969252