Make SoyExpression more flexible by getting rid of all the type specific subclasses and instead using delegation.  Also, start integrating SoyType into SoyExpression (there is more to be done here).
The motivation is that very soon we will want to do something like:
compile an expression, store it in a local/field, use that local/field as a
SoyExpression with the same type and properties as the original SoyExpression.
However, given the current implementation this is not feasible.

tl;dr; composition is preferred to inheritence.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=89974198