Soy: promote ErrorReporter to SoyFileSet.
Currently, the error reporter is owned by SoyFileSetParser, which passes it down into lower levels to collect errors during parsing. But the time has come to collect errors from the Soy codebase outside of parsing.

This CL moves the error reporter up one level, to SoyFileSet, which then injects it into SoyFileParser, and will soon inject it into the non-parsing visitors.

As a byproduct, this CL deletes ParseResult. ParseResult was a tuple containing the root of the AST and a list of parse errors, but this is no longer appropriate once errors can occur outside of parsing.

Parsing tests that used to assert that the ParseResult had no errors can now take advantage of the new behavior of SoyFileSetParserBuilder#parse, which uses the new ExplodingErrorReporter to immediately throw an AssertionError if any parse error occurs.

Parsing tests that used to inspect that actual list of parse errors (as opposed to just asserting there were no errors) can achieve the same thing by passing a custom error reporter to the SoyFileSetParserBuilder.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=90122724