Remove the type parameter from ExprRootNode and remove references to ExprRootNode from as many places as possible.
ExprRootNode implements a simple holder strategy to make it easy for various
visitor passes to replace expr trees without needing to fiddle with the
ExprHolderNode.  This on its own is a pretty reasonable strategy.  However,
outside of the SoyNode tree and various expr visitors there is really no point
in seeing this type, it just adds indirection.

So we just remove ExprRootNode from as many apis as possible and push
construction of these holders into the SoyNode constructors.  The benefits of
this are
 * remove a lot of .getChild(0) calls
 * remove worthless type parameters from ExprRootNode declarations
   The type params were only useful for propagating extra type information from
   the ExpressionParser directly to callers, but removing the ExprRootNode
   wrapper achieves the same goal.  All other declarations of those type params
   were <?>
 * (minor) move some logic out of the ExpressionParser.jj file
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=90726567