Soy JbcSrc: Add support for template parameters
Each template parameter is stored as a SoyValueProvider in a template field.  To use the field we generate code like

ResolveStatus status = field.status();
if (!status.isReady()) {
  return RenderResult.continueAfter(status.future());
}
field.resolve();

Unlike output detaches, here when we restart evaluation we start the entire expression over.  This is necessary because to jump back into the middle of an expression we would need to restore the runtime evaluation stack.  The cost of doing so (in terms of logic and implementation complexity) is considered to be negligible given that expressions should be cheap to evaluate in the first place.  This assumption may be violated for things like custom SoyFunctions which may be arbitrarily expensive.  However, if this is actually an issue for anyone trivial template changes could be made to work around it.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=90947492