Move render(AdvisingAppendable appendable, boolean isLast) up to SoyValueProvider
In order to have proper detaching on renders operations in jbcsrc, we are going
to use a different strategy than Tofu.  The issue is that in ToFu we just block
on futures, so representing a lazy content closure as a SoyValue made sense.
But for jbcsrc we need to represent all lazy closures as SoyValueProviders
(otherwise we would need to make every operation on SoyValues detachable).  For
example, a lazy content {let} in tofu.

{let $foo}... some large amount of content, potentially referencing futures {/let}

we would represent this as a StringData.LazyString, and if it got passed to a
print operator, we would be able to incrementally render it.  However, if it
got passed to a '+' operator, we would render the whole thing into a buffer and
block on a future.  This was fine because we had no better options in Tofu.

However, for jbcsrc we can do better by representing it a SoyValueProvider and
having a special render method on SoyValueProvider, for rendering partially
resolve values.

Also, since SoyValueProvider will be returning RenderResults from the new
method, return it from the status() method as well.  Maybe we should consider a
different name for RenderResult?  We could just call it Status? Outcome?
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=91028343