Extract the template name mangling to a helper class and change the generated factories to be 'inner classes'.
A benefit of this approach is that instead of using a '_' as the name factory
separator we use '$' which cannot accidentally conflict with a users template
name (since '$' isn't allowed in soy template names).  Also, if we do use '$',
then it is a best practice to actually make it an inner class so various
reflection utilities don't break.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=91114742