Soy+[]: add infrastructure and elementary unit test.
This CL creates a SoyAnalyzer that takes in a Soy AST and writes out some basic metadata in a form that [] can understand. The class structure was inspired by the js2 [] implementation, but is much simpler (for now).

The test uses Gard, a special test-only query language for [], to access the metadata written by the SoyAnalyzer.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=91589660