Make a few changes to Soy Types
1. map literals should only be interpreted as a record type if all the keys are string literals (the current check for this is insufficient, it just checks that all the keys are strings)
2. Check that all record literals only pass one instance of each key
3. Unions with UKNOWN should just be UKNOWN, 'unknown|string' should be equivalent to 'unknown'

Item #3 required a fix in the CheckCallingParamTypesVisitor to allow protos to be bassed to unknown parameters in (otherwise) strict templates.  I discussed this with talin offline and it appears that the reason to disallow passing protos to templates as unknown was to prevent users from passing protos to non-strict templates.  So instead i implemented that check.

This was refactored off of [] since this half of the cl is believed to be less risky.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=91993285