Enhance nullability analysis by allowing us to track the nullability of entire expressions instead of just variable references.
e.g.
in this template
{template .foo}
  {@param? p : int}
  {if $p}{$p}{/if}
{/template}

the type checker is smart enough to know that references to $p inside the {if} are non-null.  However, it cannot handle slightly more complex conditions
e.g.
{template .foo}
  {@param p : [a : int|null]}
  {if $p.a}{$p.a}{/if}
{/template}

In this template, "$p.a" refers to a field with the type of 'int|null', but given the condition, we know that inside the if statement, it cannot be null.  This cl enhances the type assignment analysis to cover this case.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=92208503