Automated [] rollback of changelist 92555417.
*** Reason for rollback ***

[] has detected that 1000 or more targets failed to build at [] each of which successfully built at the prior CL that affected them. To see these targets, visit: []Questions? Comments? See the URL. []

*** Original change description ***

Add a 'checkNotNull' function to soy.

checkNotNull(<someSoyExpr>) will throw a runtime exception if the argument is null and also the return type of the soy function will be the same as the argument but with all nulls removed.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=92556566