Rollforward [] which was rolled back in [] due to a bad js type parameter.  This fixes the type declaration and modifies the [] file to build the examples, which would have caught this.
*** Original change description ***

Add a 'checkNotNull' function to soy.

checkNotNull(<someSoyExpr>) will throw a runtime exception if the argument is
null and also the return type of the soy function will be the same as the
argument but with all nulls removed.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=92568608