Partial rollback of []
The problem is was that Objects.equal(a, b) is internally calls Object.equals(Object), but SoyValue has an overload of that method (SoyValue.equals(SoyValue) and subtypes of SoyValue do not consistently override Object.equals(Object) to do the right thing.

The real fix for this problem would be to get rid of SoyValue.equals(SoyValue), but for now i just inline Objects.equal so that overload selection works correctly.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=92573376