Rolling [] forward without the type enhancements for the preexisting soy nonplugin functions. It turns out that user templates depend on the 'unknown' types that they generate.
*** Original change description ***
Add a 'checkNotNull' function to soy.

checkNotNull(<someSoyExpr>) will throw a runtime exception if the argument is
null and also the return type of the soy function will be the same as the
argument but with all nulls removed.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=92597462