Soy: propagate SoyErrors through SoyJsSrcResource.
The main top-level entry point for compiling Soy to JS is SoyToJsSrcCompilerForGoogle, which is a standalone java_binary. It has logic for pretty-printing SoyErrors and exiting with an appropriate error code.

There is an older Soy->JS compilation entry point: SoyToJsSrcResource. There are ~10 users of this in [], and it has not been open-sourced. A user of this class recently complained about being unable to inspect SoyErrors.

This CL changes SoyJsSrcResource to aggregate the SoyErrors and re-throw them as suppressed exceptions under a single SoySyntaxException. Users of SoyJsSrcResource can iterate over these suppressed exceptions and take appropriate action.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=93530538