Run contextual autoescaper on private templates, too.
It turns out there are some private templates that are being called either
correctly from templates in other files, or incorrectly from Javascript.  Not
running the contextualizer on these private templates means we have some random
templates that should be contextual, but are not contextualized.

We discovered this fact when trying to submit [] which attempted to
add a check to PerformAutoescapeVisitor that ensured that every print node in a
contextual or strict template has an explicit escaping directive, and that the
"default escaping mode" was only used for noncontextual templates.  Because of
that discovery we realized we needed to explicitly contextualize these weird
private templates.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=94291736