Soy ContentSecurityPolicyPass: make nonce injection more conservative.
Currently, the csp_nonce param is injected as nullable, but not null safe, causing gencode like this:

{if $ij.csp_nonce}
...

If the caller (Tofu, for example), forgets to pass in an injected data record, the result is the Soy equivalent of an NPE.

This is bad. The decision to use CSP nonces should be independent from how injected data is configured. If CSP nonces are enabled but no injected data is available, Soy should degrade gracefully and not attempt to dereference the param.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=94579873