Fix call-site escaping in html nospace context.
The problem was we failed to take into account the epsilon transition,
and it was failing silently with an empty list of escaping modes, which
really means "error" but wasn't being checked.

Added a TODO to check it (I was already planning on making this change
but with the intent of improving error messages, had no idea about this
bug until Emmanuel pointed it out).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=95860683