Forbid URI syntax in Soy where the scheme could be set by an attacker.
This ensures more safety in Soy, but does enforce some additional rigor in URI
syntax.

As part of this, I refactored the contextual autoescaper to remove the ERROR
state, because it got in the way of good error reporting.  Instead of an ERROR
state, it makes more sense to throw errors as soon as possible, where the best
error message can be provided, and allow the caller to catch the exception and
associate the meta info.  The exception is Context.union(), where it's not
really possible to make an informative error message anyway, in which case I
chose null.

Design doc:
https://docs.google.com/document/d/1VNnvK8oVp7xCiB8M9HjJ4-MlrQ38UrkI6N4DiaW98zU/edit
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96019197