Soy Jbcsrc:  introduce a common class for classwriting in jbcsrc.
This is a minor cleanup to reduce duplication/increase consistency and also to provide a place to override ClassWriter.getCommonSuperclass.

ClassWriter.getCommonSuperclass, is a method used in frame merge calculations to compute frames.  For example, imagine the java expression:

someCondition() ? <left> : <right>

to calculate the type of this expression we need to know the least common supertype of left and right.  internally asm needs to do this calculation in order to generate stack frames and it does it via (among other things) getCommonSuperclass.  The default implementation works by loading the classes via reflection and using methods on the Class object.  This does not work if one of the types in the comparison is one of our generated classes.  I'm still not entirely sure what the best approach is to implementing this method so there will be some followup work when i understand the situtaion better.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96038751