Fix clone for LocalVar.
LocalVar and LocalVarNode form a cycle in the AST tree (LocalVarNodes have local var fields, and LocalVars have declaringNodes()).  This is useful for being able to determine more information about a local var but it makes .clone() more difficult to implement.  The issue is that the current version of LocalVar.clone(), doesn't clone  its declaringNode and instead just retains the original reference, which means that in practice the AST cycles are broken and cloned trees retain references to the tree from which they were cloned.  In particular, this is causing a bug in jbcsrc where we can't determine where a local var is defined because we start traversing up from the declaringNode() and wind up in a detached AST subtree!

The solution is just to traverse the tree and fix all the local var vardefns after cloning.  A better solution would be to stop using the Object.clone() api and instead use a custom Node.clone(CloneState) apis then the CloneState object could track the additional state needed to resolve the cycles.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96160181