Replace the use of Object.clone() with a new Node.copy(CopyState) method with nearly the same rules.
This resolves some weirdness that exists purely because clone() is defined on
Object and Node. Also, by adding an extra parameter to this method we can
efficiently (and correctly) reconstruct the full AST in one pass and get rid of
the multipass helpers in SoytreeUtils.  This cl doesn't actually do that, but
it lays the (boring) groundwork.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96311412