[]: RFC: implement CompilationAnalyzer for Soy.
According to [] a requirement for integrating a new language with [] is to implement the CompilationAnalyzer [] service. This CL does that (SoyCompilationAnalyzer), and exposes it via a new Java binary, soy_service_analyzer.

In order to continue using SoyAnalyzer in unit tests, this CL introduces a SoyAnalysisDetails class that abstracts away the details of how inputs are provided.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96410348