Modify the TODO in SoyClassWriter now that i have a better idea about what is going on.
Basically, given a template like:
{@param list :list<string>}
{foreach $item in $list}...{/foreach}
{let $foo : 'a' /}
{$foo}
I will generate code that stores both $list and $foo in a local variable.  It happens to be the case that $foo and $item have non-overlapping scopes so I take advantage of this and store them both in local variable slot #3.

This means that at various points slot 3 holds List or a soyValueProvider so ASM needs to know their least common super type in order to generate a valid stack map frame for slot #3.  This is kind of unfortunate, because it should be possible to tell ASM that the first use of #3 doesn't overlap with the second usage.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97089362