Soy jbcsrc: fix code generation for non-trivial for loops
The fix is sadly quite subtle, though i have introduced tests that should prevent regression.  The issue is that the order of operations in SoyNodeCompiler.calculateRangeArgs needs to be the same as the order of the generated code.  This is because compiling expressions to bytecode changes the DetachState object and then later detach points will assume that all prior generated detaches have already been added to the bytecode, if this invariant isn't maintained we will end up generating code which uses undefined variables which is exactly what happened here.  This would only happen in {for} loops with non constant range arguments and it turns out there were no tests for that!

To fix i simply moved more logic into calculateRangeArgs and reordered some
operations to maintain invariants.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97132619