Add support for the css node with an API similar to the goog.getCssName functions.
This relies on a module level object to contain the css mappings. This isn't
really ideal, but the only alternative API would be a configurable function to
provide the css mapping at compile time. That feels too heavy to require for
css mappings, so I kept it simple. I'll see how this feels with the YT
implementation and possibly revisit in a future release.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97794816