Soy: improve internal DeclInfo API to pave way for source locations.
Currently, TemplateNodeBuilder reports SoySyntaxExceptions arising from malformed @param declarations. I'd like to change this to use SoyErrors, but to do so, I need to make DeclInfo objects aware of the SourceLocations that they come from.

This CL adds SourceLocations to DeclInfo objects, and makes a number of improvements to the API along the way. For example, the DeclInfo constructor currently takes an arbitrary string as the logical "kind" of the parameter, even though there are only two choices (@param and @inject). This CL replaces that param with an enum.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97914011