Allow data access expressions (i.e. the '.' and '[..]' operators), use arbitrary primary expressions for base their base expressions.
This fixes a longstanding todo with soy's expression grammar, previously you
could only use other 'data expressions' as the base expression for a data
expression.  so things like:

['1','2'][0]
['a':'b']['a']
['a':'b'].a
($a ? $b : $c).field

all become valid soy expressions with this change.

To implement the change, I ultimately rewrote all of Soy's operator parsing to
use a more traditional (if verbose) strategy.  The problem with the previous
version was that it was overly tied to the Operator enum which made fitting '.'
and '?.' into it difficult.  I do think this version is easier to understand,
as well.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=98671302