soy jbcsrc: Reduce the number of places we reference the soy type system.
The soy type system has a number of places where it gets out of touch with
reality and/or implements slightly looser semantics.  This have been the cause
of much pain.  So instead we limit the number of places that we reference it
and instead rely on our own understanding of the operators.

With this cl we (pretty much) only use the type system for params, data
accesses and literals.  There is probably more to remove but this is the bulk
of it.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=100084684