Get some better error messages out of the type parser.
Instead of:
  Exception in thread "main" com.google.template.soy.types.parse.TokenMgrError: Lexical error at line 1, column 6.  Encountered: "!" (33), after : ""
          at com.google.template.soy.types.parse.TypeParserTokenManager.getNextToken(TypeParserTokenManager.java:353)
          at com.google.template.soy.types.parse.TypeParser.jj_ntk_f(TypeParser.java:419)
....

We now have:
In file soy/integrationtest/basic_test.soy:14:3: parse error at '!' : expected [, ?, list, map, or an identifier
  {@param p : list<!string>}
  ^

Which is better.  However, source location data is still inaccurate and it is not easy to fix.  So that will have to wait for rewriting the TemplateParser.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=100581747