Remove v1 inference rules based on namespace.
Currently there are 2 rules:
1. no namespace in the file => v1
2. template with fully qualified name => v1

With this change we only allow templates to be declared with fully qualified names if it is also marked with deprecatedV1="true" and we just drop the requirement that files have namespaces.  Instead, we rely on a preexisting check that templates with relative names can only be defined in files with a declared namespace.  This should be just as good, but has the strange result that the empty string is now a valid SoyV2 file and files that contain only deltemplates no longer require namespaces (this seems reasonable to me).

As per usual, the biggest pain is in the unit tests, which are all written in v1 syntax.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=101093962