Disallow partial tag names in Soy.
After this, a tag name is either fully specified or not specified at all.

Allowed:

  <{if $x}div{else}span{/if}>
  {if $x}<div{else}<span{/if}>

Banned:

  <s{if 1}cript{/if}>Smuggled script tag</script>
  <h{$x}>Variable header number</h{$x}>

This fixes a long-standing hole in Soy's element name detector.

This also has the benefit that we can determine the element type early. Soy
used to eagerly determine the ElementType for special tags, but delay for
normal tags.  For example:

  <script{if} -- at this point Soy already knows it's a script
  <div{if} -- Soy would delay determining what type it was

This made it hard to add support for:

  {if $iframe}<iframe{else}<img{/if}

which one team did. In the img branch, Soy already knew it was an img (since
img is special), but for iframe it did not know what type it was.  It was
fixable by guessing, but it really just makes sense to choose the tag name
quickly.

Note I also uncovered some hidden gems. Soy wasn't distinguishing much from
start and end tags, so I added an extra state to account for this. In doing
so I discovered that Soy was being a little lax in parsing, so I banned the
following:

  </3 is invalid HTML, since </ can only be followed by a valid tag name
  </script>, </style>, etc will be an error if we didn't see the open-tag
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=101596097