Refactor Soy's Context.union() and add tests.
The old version of union() had a bunch of special cases for when contexts could
be unioned or not.  But this wasn't really scalable, since every time a new
field is added to Context, there were a bunch of places where that field might
end up ignored. For example, some of the checks would see if some properties
matched, and if they did, just returned either the first or second context.

This version instead looks for specific differences, and "reconciles" the
difference. It looks more like:

  - Do "a" and "b" have different states?
  - If so, see if we can change the state of a or b in allowable ways.
  - Repeat for other fields (e.g. if they have different URI types or JS-following-slash types)
  - Check to see if a and b are now exactly equal.

As part of this:

  - Added unit test for Context
  - Moved the parsing function into Context itself for easy use in tests.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=101616862