Roll forward [] with fix for Tofu function adaptation.
[] attempted to consolidate SoyFunction lookup in a single Map<String, SoyFunction>, which ResolveFunctionsVisitor used to decorate the AST.

That CL missed a significant side channel for SoyFunction injections: the @Tofu Map<String, SoyJavaFunction> binding. That binding was provided by TofuModule and contained logic to adapt legacy SoyFunctions (specifically, SoyJavaRuntimeFunction and SoyTofuFunction) into the canonical interface (SoyJavaFunction).

The CL deleted a @Tofu binding; it assumed (wrongly) that all the SoyJavaFunctions available through the @Tofu binding were also available in the top-level Map<String, SoyFunction>. As a result of this incorrect assumption, SoyJavaFunction lookup would fail in the EvalVisitor, and the CL had to be rolled back.

This new CL rolls forward the original one with a fix. Adaptation of SoyJavaRuntimeFunctions and SoyTofuFunctions is now done at the top level, in SoyFileSet, when a Tofu compilation is requested.

As a result, the @Tofu Map<String, SoyJavaFunction> binding is unused and therefore removed. (The other use of @Tofu, to provide SoyJavaPrintDirectives, is unaffected.)
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=102347389